update-project

+++ cart.php
@@ -62,28 +62,27 @@ if (isset($_GET['delete_all'])) {
     <div class="box-container">
       <?php
       $grand_total = 0;
-      $select_cart = mysqli_query($conn, "SELECT * FROM `cart`
-      WHERE user_id = '$user_id'") or die('query failed');
+      $select_cart = mysqli_query($conn, "SELECT * FROM `cart`, `products`
+      WHERE user_id = '$user_id' AND product_id = products.id") or die('query failed');
       if (mysqli_num_rows($select_cart) > 0) {
         while ($fetch_cart = mysqli_fetch_assoc($select_cart)) {
       ?>
 
-      <div class="box">
-        <a href="cart.php?delete=<?php echo $fetch_cart['id'] ?>" class="fas fa-times"
-          onclick="return confirm('delete this form cart?');"></a>
-        <img src="uploaded_img/<?php echo $fetch_cart['image']; ?>" alt="">
-        <div class="name"><?php echo $fetch_cart['name']; ?></div>
-        <div class="price">$<?php echo $fetch_cart['price']; ?>/-</div>
-        <form action="" method="post">
-          <input type="hidden" name="cart_id" value="<?php echo $fetch_cart['id']; ?>">
-          <input type="number" name="cart_quantity" min="1" value="<?php echo $fetch_cart['quantity']; ?>">
-          <input type="submit" name="update_cart" value="update" class="option-btn">
-        </form>
-        <div class="sub-total">sub total :
-          <span>$<?php echo $sub_total =
+          <div class="box">
+            <a href="cart.php?delete=<?php echo $fetch_cart['id'] ?>" class="fas fa-times" onclick="return confirm('delete this form cart?');"></a>
+            <img src="uploaded_img/<?php echo $fetch_cart['image']; ?>" alt="">
+            <div class="name"><?php echo $fetch_cart['name']; ?></div>
+            <div class="price">$<?php echo $fetch_cart['price']; ?>/-</div>
+            <form action="" method="post">
+              <input type="hidden" name="cart_id" value="<?php echo $fetch_cart['id']; ?>">
+              <input type="number" name="cart_quantity" min="1" value="<?php echo $fetch_cart['quantity']; ?>">
+              <input type="submit" name="update_cart" value="update" class="option-btn">
+            </form>
+            <div class="sub-total">sub total :
+              <span>$<?php echo $sub_total =
                         ($fetch_cart['quantity'] * $fetch_cart['price']);  ?>/-</span>
-        </div>
-      </div>
+            </div>
+          </div>
       <?php
           $grand_total += $sub_total;
         }
@@ -95,8 +94,7 @@ if (isset($_GET['delete_all'])) {
 
     <div style="margin-top: 2rem; text-align: center;">
       <a href="cart.php?delete_all" class="delete-btn <?php echo ($grand_total > 1) ?
-                                            '' : 'disabled'; ?>"
-        onclick="return confirm('delete all from cart?');">delete all</a>
+                                                        '' : 'disabled'; ?>" onclick="return confirm('delete all from cart?');">delete all</a>
     </div>
 
     <div class="cart-total">
